[PiP on Desktop] Don't reparent PiP task to Desk if mid-Recents
As a temporary workaround for Desktop PiP mid-Recents, we previously
made it so that PiP will always expand to fullscreen mid-Recents in
Ibcdc47054aea52092341e55e2ee435543db58d5d.

This CL expands on the workaround to also bypass reparenting the task to
the Desk when multiple_desktops_backend flag is enabled. Note that this
fix is temporary to avoid getting into a bad state -- a more
comprehensive fix for PiP in Recents will be in a future fix.

Bug: 419875246
Test: atest DesktopPipTransitionControllerTest
Test: Manual - enter PiP on Desktop, swipe up to Recents and expand PiP.
Verify PiP expands to fullscreen
Flag: com.android.window.flags.enable_desktop_windowing_pip

Change-Id: Iba7161139ac44c3225344a3c3a2a22fd0a09f892