Enter device on non-bypass face auth on transition to altBouncer
If the device is transitioning to the alternateBouncer when
face auth succeeds, make sure to enter the device and
dismiss the lockscreen upon face auth success.

Previously, if the device was in the middle of a transition
to the alternate bouncer when face auth succeeded,
the user would be on an unlocked lock screen and
needed to swipe up  or longpress the device entry
unlock icon to enter the device.

Fixes: 413626364
Test: atest BiometricUnlockControllerTest
Test: Enroll udfps and face auth, on the lockscreen quickly
tap on a notification and face auth successfully multiple times;
observe each time on face auth success, user enters the device
Flag: EXEMPT bugfix

Change-Id: I8d2b0189304d44616f98f89c4e1540b0ba99b870