Improve IME layering/input/control target dumping
This starts dumping the IME layering, input and control targets in the
DisplayContent, and removes the dump from WindowManagerService (which
was originally added in [1]). This brings the dumped information in the
same place as its declaration and update logic.

Additionally improves the format of SoftInputShowHideHistory, by
updating the legacy imeTargetNameFromWm name, and by also dumping the
IME input target.

Lastly, this adds a toString for the RemoteInsetsControlTarget, so we
can see which display it corresponds to, and what requested visibile
types it has.

Flag: EXEMPT log fix
Bug: 281029564
Test: verifying output of adb shell dumpsys window
  and adb shell dumpsys input_method

  [1]: I9aa934961fb3975bd2af18599b5a2884387b5007

Change-Id: I3e1ea6836638ee50575bea9bd34c301ce608f5a6