[47/n] Simplify Letterbox Surface lifecycle
Removing the Letterbox surfaces created when a letterboxed activity
appears in a Task can be done when the Task itself is removed.
In the existing LetterboxLifecycleController is much easier to
just handle creation, visibility and resizing of the letterbox
surfaces.

Also moving the resizing of the Letterbox surfaces later in the
lifecycle logic, allows to handle the case when the surfaces have size 0.
In that case the Crop would not happen causing an unwanted behaviour
of a surface filling the task completely which is obvious in case
of multiple surfaces for a transparent activity.

Flag: com.android.window.flags.app_compat_refactoring
Bug: 409951687
Test: atest WMShellUnitTests:LetterboxLifecycleControllerImplTest
Test: atest WMShellUnitTests:MigrationLetterboxTransitionObserverTest
Test: atest WMShellUnitTests:DelegateLetterboxTransitionObserverTest

Change-Id: If411b45f25aec74447e34753cb3997a22b8352f1