Replace FixedInstant by FixedDate + FixedTime
Using LocalDate and LocalTime removes the need for TimeZone shenanigans, and also for choosing whether to display date or time. The date+time combination is not a good fit for the limited space of metric anyway.

Also, remove FORMAT_CHRONOMETER_MINUTES and adjust naming of remaining options.

Bug: 415827681
Test: atest FrameworkCoreTests:NotificationMetricStyleTest
Flag: android.app.api_metric_style
Change-Id: I3b70ea6acd621ba9039b7cb10d26eea79d5cc771