Validate user on biometric auth and keyguard done
On user switching, it is possible to start a chain of events
that validate the prior user before user switching is complete,
notably on biomtetric auth success and the tryKeyguarDone paths.
The latter is a runnable that can be scheduled after dismiss
animations run, but if the user was changed in the middle, it
is possible to dismiss the prior keyguard.

Revalidate the user at multiple points, and reset the keyguard
when the user id mismatch is detected.

Bug: 413380719
Test: atest KeyguardViewMediatorTest
Flag: EXEMPT bugfix
Merged-In: I70efe8704931e096698d4a775ee487bb5241d7f1
Change-Id: I70efe8704931e096698d4a775ee487bb5241d7f1
(cherry picked from commit 67c08855f57a2abec6984471ef2b47a8da87b3c6)