Prevent WallpaperActivity from being added to repo
In some cases, DesktopWallpaperActivity is being launched in freeform.
To prevent the device from being unusable when that happens, prevent
adding it to the DesktopRepository. This CL should be reverted when the
root cause for the above is found and fixed.

Bug: 420917959
Test: atest DesktopTaskChangeListener
Flag: com.android.window.flags.enable_windowing_transition_handlers_observers
Change-Id: Id81587baefa426b6a05ad849f36d991c7f0669ec