Stop depending on transitions for tasks moving displays.
On switching users, getDeskIdForTask returns null for tasks that are
in the newly activated desk due to the repositories not being updated
before the onTransitionReady signal.

This CL stops depending on transitions to break tiling when a task is
moved across displays, and instead breaks tiling when the code to move
the task runs.

Flag: com.android.window.flags.enable_tile_resizing
Test: Unit tests and on device testing
Bug: 416304871
Change-Id: Ib83bc321cb1d578997d6c132c7d7ac14186384f0