System UI: Update rendering of RegionBox to place knobs on top of border
We have to make sure that the borders of the rectangle is rendered
before the knobs. to do this, an invisible rectangle is drawn with
contains the area for the border and the knobs, then the border and
knobs are rendered within the same block to ensure of this consistency.

Screenshot: http://screenshot/4dPbKbv8HYSnjBk
The red background is for visibility only and is not included in this
CL.

Bug: 417533606
Test: Manual
Flag: com.android.systemui.desktop_screen_capture
Change-Id: I3e2462b89c1ee51d4d68a2c5c2baf38c1dfcad5d