Fix OverlayRemountedTest for HSUM
Devices with multiple users (such as HSUM) cause the output of `cmd
overlay dump isenabled` to have multiple lines, one for each user. The
tests fail trying to parse the output. The fix here is to enable the
overlay and lookup the resource value only for the current user.

Also removed some unused code in SystemPreparer.

Fixes: 396551034
Fixes: 406547979
Fixes: 414526588
Test: atest OverlayRemountedTest on HSUM device
Flag: EXEMPT bugfix

Change-Id: I8ff1b8e0884b5a7aaf8b4bed05fcc1e311c84bba