[Reland][Flexiglass] Use BC25 brightness bar in the QS scene
The landing of the previous CL ag/33784459 caused crashes due
to an irrelevant bug which has been fixed by ag/33896778.
This CL is to reland it.

This CL also ensures ShadeHeaderViewModelKosmos has a meaningful
mobileIconsViewModelKairos factory.

The previous CL's description:

This CL uses the BC25 brightness bar in the QuickSettingsScene.

This CL also adds a compose test to verify the view hierarchy of
the QuickSettingsScene.

Bug: 420960164
Flag: com.android.systemui.scene_container
Test: added a compose test
Change-Id: I1d4e90469ca739800f5937461aa8b19147b7566f