Screen Capture: Setup ViewModel and State
Basic setup for the ScreenCaptureViewModel which holds the state for the
pre-capture UI. The state for the toolbar radio buttons is propagated
down to the composables. Also, the RadioButtonGroup API was updated to
make it easier to configure the click action for each button.

Bug: 422855266
Test: atest ScreenCaptureViewModelTest
Flag: com.android.systemui.desktop_screen_capture
Change-Id: I2ec944f243e5e2eb45c6738a6a69a2e12e8f55ab