Fix LauncherProxyServiceTest
Test was failing on specific builds due to a HSUM issue. The userId didn't match the calling id in the test, which stopped the expected method from being called. Fix is to just mock the userId to always match the calling userId

Bug: 362719719
Bug: 417734966
Bug: 419057255
Test: LauncherProxyServiceTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I5088f6d19a532438c5064367760657da74bc2d9f