Revert^4 "Migrate usages of ViewConfiguration static methods"
This reverts commit 725715c8f56736e72f3cb2efa477b343a06ac7a5.

Reason for revert: Fixes root cause of b/422829787, by avoiding the  caching of long press timeout as a member during creation time in GestureDetector and various other widget classes. Instead, they keep ViewConfiguration instance as a member and fetch the long press value from it when needed. The View class exposes getLongPressTimeoutMillis as a protected (albeit hidden) method, so that subclasses within the framework can directly call that, instead of keeping their own instance of ViewConfiguration.

Bug: 370720522
Change-Id: If36351e1784d4b3bbaf8b8614d8099ad8b0329fe