Force screenshot sound if camera open
In certain locales, it's required for the camera to make a sound when a
photo is taken, even if the phone is otherwise on silent. Originally, we
turned off this feature for screenshots (since most screenshots are just
of the screen and don't pose a privacy concern in the same way).
However, it's possible to take a screenshot *of the camera*.

This change forces the shutter sound on if it's required for the camera
and the camera is currently open.

Bug: 411503333
Fix: 411503333
Test: manual (by hardcoding the shutter sound to be forced on)
Flag: com.android.systemui.screenshot_force_shutter_sound
Change-Id: Ie4e16ce3846369b5a5c6bd24a24139eaf420d2e5