Glanceable hub is blurred when swiping down shade a little
The issue occurs when swiping down shade over hub a little. If
handleExternalTouch triggers onInterceptTouch again and it returns
false, but a prior call to shouldInterceptTouchEventInternal was
intercepted by the DragDown helper with dragging in progress, ACTION_UP
will not be processed, causing the shade to get stuck.

Fix is to avoid calling onInterceptTouchEvent when DragDown helper has
already intercepted and dragging is in progress.

Bug: 412261838
Test: pull down shade over hub slightly and fully, across various distances
Test: atest NotificationShadeWindowViewControllerTest
Flag: com.android.systemui.hub_blurred_by_shade_fix
Change-Id: Ieef2ff8717c02721297c82dd2d504b38269ecbc1