Revert^2 "Update ServiceRecord's appInfo on overlay changes"
This reverts commit 8dd6ed0a612282308612882e9f778e55c00ef48e.

Reason for revert: Reapplying with the fix
------
Original message:

Update ServiceRecord's appInfo on overlay changes

ApplicationInfo in ServiceRecord is rarely used, so the code
used to not update them for runtime overlay changes and there
were no known issues with that. Turns out, there's a specific
case when it does matter:

An app has a service that's auto-restarting, and an activity
that starts the service. If any overlay changes happen after the
service start, we never update the service record. Then, if the
activity crashes, we restart just the service using it's (stale)
application info, and it doesn't get the latest overlay paths.

This change applies the new ApplicationInfo when
OverlayManagerService sends an update also to all running
services in the process

Bug: 414686685
Test: manual with the CL attached to the bug
Flag: android.content.res.update_ams_service_appinfo
Change-Id: I69727f301ca67bc2a21d9511464dc3567515f48e