[SwipeHelper] remove not swiped out notification from swiped out list
It is expected that Swiped out view is moved to transientview first and after that swipe out animation runs. However, this animation is not part of NSSL animations and it runs independently.
This animation starts before NSSL removes this notification and get cancelled due to some updates. In this edge case, this view is moved to transientview and remain on the shade.

This CL removes this notification from swiped out list when it is not swiped out and leaves removal responsibility to NSSL.

Fixes: 415279408
Test: Post  a notification with notify with delay and have many updates. During updates try to swipe out quickly.
Flag: EXEMPT bug fix
Change-Id: I18a8b95501fd93f41571dd71ba5a20ab24a87c9c