Always scan all packages on user unlock
Modifies ShortcutService.checkPackageChanges so that it will always scan
all packages for the user on unlock.

Currently, checkPackageChanges will only rescan packages that have a
PackageInfo.lastUpdateTime > ShortcutUser.lastAppScanTime.

This causes an issue if the service was unable to successfully write all
the scanned shortcut info, as it will not rescan these packages when
the user is unlocked the next time.

Also, since the last user unlock, aconfig flags may have been flipped
that cause new shortcuts to be available, even if the
PackageInfo.lastUpdateTime hasn't changed.

This change also re-enables some unit tests that exercise
checkPackageChanges, as well as ShortcutService saving and loading.

Bug: 417712196
Test: ShortcutManagerTest1
Flag: EXEMPT bugfix
Change-Id: I1944f5996fd3da9cea41c5ed0943f7cdd0cf076f