Demote exception to log message
For some race condition it might happen that the parent surface
is not present when creating a CompatUI components.
This might happens when the parent surface is removed and some
steps in the compat ui component lifecycle tries to add the component
again.

In that case it's ok to ignore the operation and just avoid to add the
child Surface.

Flag: EXEMPT Refactoring
Fix: 408185221
Test: atest WMShellUnitTests:CompatUIWindowManagerTest

Change-Id: Id0c388d8488fa4b44a67181f65d0a8d9a3d60d95