[SB] Add flag to always use color sampling for status bar.
In Idf6d34840ee2bc9bca1ca0d0c44ad930ed6989cb, we added the ability to
use color sampling for the status bar when the dark theme accessibility
feature is enabled (see b/379760792). But, we've been wanting to do
color sampling in the status bar for all users all the time, so that we
don't have apps incorrectly setting the status bar color and making it
invisible.

This CL adds a flag that will enable it at all times. There's still a
lot of testing with various apps & performance testing to be done, but
adding the flag will let us start that sort of testing.

Bug: 365120736
Flag: com.android.systemui.status_bar_always_use_region_sampling
Test: atest StatusBarRegionSamplingInteractorTest
Test: create demo app that can set area behind status bar to any color
-> verify status bar automatically changes between light & dark icons
based on color (see videos in bug)

Change-Id: Ie4add4e96b500718252dbb7d9baff0b5cb012c9e