fix: Clear the icon dark area when all icons should be light
Issue: sometimes the battery icon is incorrectly dark when all other
icons are light.

This change addresses that problem for the specific scenario of "all
icons should be light" by ensuring that we clear out leftover
DarkIconDispatcherImpl#mTintAreas.

This change does NOT address that problem for the scenario when the
icons are supposed to be half light and half dark (but battery alone is
still broken) e.g. on a large screen device in split screen.
That case still needs further investigation.

Bug: 423390143
Bug: 379760792
Flag: EXEMPT small and low risk bugfix
Test: LightBarControllerTest
Test: Observe battery icon matches other icons, when all icons are
      supposed to be light.
Change-Id: I2a0f55d5d9ce4556509023dd54193b7ba5c12bea