Properly gate reconnect events.
Fixes an issue where default desks are not created when reconnecting a
display with the reconnect flag off by fixing the flag gating of the
feature. Now, with the flag off, displays are not preserved and
restoreDisplay is not called, ensuring default desks still get created.

Bug: 424065356
Test: Manual; attempt reconnecting with flag on and off, ensure default
desk is not prevented with flag off
Flag: com.android.window.flags.enable_display_reconnect_interaction

Change-Id: I3b952757b9c24c85240cc55f0561c5a556f21236