Move `sanitizeStateParams` ahead of Bundle modifications
Currently, if Bundle has to check for `hasBinders`, the bundle has to be in parcelled state. By adding a remove(key), Bundle was unparcelled and it led to crash for that reason.

Added a TODO to followup with more robust exception handling in this scenario to not crash.

Change-Id: I2ca25503944c8beaef103f419a9472468cee7988
Flag: EXEMPT bugfix
Bug: 372658837