Decrease UserLifecycleTests.mUserSwitchTimeoutMs to 10 seconds.
In UserLifecycleTests, user switch timeouts in 30 seconds, and
these tests set UserSwitchObservers' timeout to 100 seconds for
the test session. But this doesn't make sense and causes the
logged reason behind the timeout to be less granular.

This CL changes UserSwitchObservers' timeout to 1/3 of the user
switch timeout, because UserSwitchObservers wait on that timeout
twice, in both onBeforeUserSwitching and onUserSwitching events.

Bug: 416164091
Test: atest UserLifecycleTests
Flag: TEST_ONLY
Change-Id: Ia7a5dad6e6571440fff9913f06e8a557ff15ea34