Direct set activity type to requested config when creating task
The simple setter is >10 times faster.

The setActivityType will invoke onConfigurationChanged that runs
the config resolving logic. It is unnecessary for a task which is
not attached yet, because when the task is added to a parent,
the onConfigurationChanged will run again to handle latest state.

Bug: 297502610
Flag: EXEMPT reduce duplicated invocation
Test: TaskTests
Change-Id: Ib11bbe19ab025665485b0448d24e6f1ad625f3c5